swarm/storage: Use GenerateRandomChunk in mputChunks too

+++ swarm/storage/ldbstore_test.go
@@ -74,7 +74,7 @@ func (db *testDbStore) close() {
 	}
 }
 
-func testDbStoreRandom(n int, processors int, chunksize int, mock bool, t *testing.T) {
+func testDbStoreRandom(n int, processors int, chunksize int64, mock bool, t *testing.T) {
 	db, err := newTestDbStore(mock, true)
 	if err != nil {
 		t.Fatalf("init dbStore failed: %v", err)
@@ -83,7 +83,7 @@ func testDbStoreRandom(n int, processors int, chunksize int, mock bool, t *testi
 	testStoreRandom(db, processors, n, chunksize, t)
 }
 
-func testDbStoreCorrect(n int, processors int, chunksize int, mock bool, t *testing.T) {
+func testDbStoreCorrect(n int, processors int, chunksize int64, mock bool, t *testing.T) {
 	db, err := newTestDbStore(mock, false)
 	if err != nil {
 		t.Fatalf("init dbStore failed: %v", err)
@@ -221,7 +221,7 @@ func TestMockIterator(t *testing.T) {
 	testIterator(t, true)
 }
 
-func benchmarkDbStorePut(n int, processors int, chunksize int, mock bool, b *testing.B) {
+func benchmarkDbStorePut(n int, processors int, chunksize int64, mock bool, b *testing.B) {
 	db, err := newTestDbStore(mock, true)
 	if err != nil {
 		b.Fatalf("init dbStore failed: %v", err)
@@ -230,7 +230,7 @@ func benchmarkDbStorePut(n int, processors int, chunksize int, mock bool, b *tes
 	benchmarkStorePut(db, processors, n, chunksize, b)
 }
 
-func benchmarkDbStoreGet(n int, processors int, chunksize int, mock bool, b *testing.B) {
+func benchmarkDbStoreGet(n int, processors int, chunksize int64, mock bool, b *testing.B) {
 	db, err := newTestDbStore(mock, true)
 	if err != nil {
 		b.Fatalf("init dbStore failed: %v", err)
